Fixed UI changes READ NOTES
Changed the bottom central click to hit next on the Quest/Missions/whatever you want to call it -window right before the friend selection. This MAY make in incompatible with some setups of FFBE over Amazon, because of the "Amazon Gaming Circle" popup when launching the app. There, for whatever reason, are different positions this banner shows up -but it's consistent, so if it works once it always works.